refactor: queryFn 분리
